+++ src/sql/parts/notebook/models/notebookContexts.ts
@@ -8,7 +8,7 @@
 import { nb } from 'sqlops';
 
 import { localize } from 'vs/nls';
-import { IDefaultConnection, notebookConstants, INotebookModelOptions } from 'sql/parts/notebook/models/modelInterfaces';
+import { IDefaultConnection, notebookConstants } from 'sql/parts/notebook/models/modelInterfaces';
 import { IConnectionManagementService } from 'sql/platform/connection/common/connectionManagement';
 import { ConnectionProfile } from 'sql/platform/connection/common/connectionProfile';
 import { IConnectionProfile } from 'sql/platform/connection/common/interfaces';