+++ src/sql/platform/restore/common/restoreServiceImpl.ts
@@ -13,14 +13,14 @@ import * as sqlops from 'sqlops';
 
 import { ICapabilitiesService } from 'sql/platform/capabilities/common/capabilitiesService';
 import { IRestoreService, IRestoreDialogController, TaskExecutionMode } from 'sql/platform/restore/common/restoreService';
-import { OptionsDialog } from 'sql/base/browser/ui/modal/optionsDialog';
+import { OptionsDialog } from 'sql/workbench/browser/modal/optionsDialog';
 import { RestoreDialog } from 'sql/parts/disasterRecovery/restore/restoreDialog';
 import * as ConnectionConstants from 'sql/platform/connection/common/constants';
 import { MssqlRestoreInfo } from 'sql/platform/restore/common/mssqlRestoreInfo';
 import { IConnectionProfile } from 'sql/platform/connection/common/interfaces';
 import { ProviderConnectionInfo } from 'sql/platform/connection/common/providerConnectionInfo';
 import * as Utils from 'sql/platform/connection/common/utils';
-import { IObjectExplorerService } from 'sql/parts/objectExplorer/common/objectExplorerService';
+import { IObjectExplorerService } from 'sql/workbench/services/objectExplorer/common/objectExplorerService';
 import { ITaskService } from 'sql/platform/taskHistory/common/taskService';
 import { TaskStatus, TaskNode } from 'sql/parts/taskHistory/common/taskNode';
 import * as Constants from 'sql/common/constants';